+++ types/interfaces.d.ts
@@ -2,7 +2,7 @@ import { SonicBoomOpts } from 'sonic-boom';
 
 export interface StructuredTransformationConfig extends SonicBoomOpts {
   /**
-   * pino ignore keys, filters Telemetry properties
+   * pino ignore keys, optional list of pino ignore keys
    * @default {string[]} hostname pid, level, time, msg
    */
   ignoreKeys?: string[];