+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## [0.0.3] - 2025-04-03
+
+- map error stack to textPayload
+- ignore mixed in properties that are null or undefined
+- start documenting api
+
 ## [0.0.2] - 2025-03-31
 
 - map error stack to source location